pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "geo-cli"
version = "2.0.0"
description = "Generative Engine Optimization CLI — AI시대의 SEO 분석 도구"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "websockets>=12.0",
    "playwright>=1.40.0",
    "pandas>=2.0.0",
]
legacy = [
    "streamlit>=1.31.0",
    "playwright>=1.40.0",
    "pandas>=2.0.0",
]

[project.scripts]
geo = "geo_cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["geo_cli*", "backend*"]
